requirements.txt
python-binance
python-dotenv
questionary
Flask 